requirements.txt
pyhocon
imgaug
flask>=1.0.0
scikit-learn
cython
numpy
cffi
opencv-python
scipy
easydict
matplotlib
Pillow
pyyaml
visdom
bs4
html5lib
ninja
torch
torchvision
